About to combine Cell and CellState

+++ ArduinoProjects/Suduino_04/Debug.h
@@ -1,18 +1,18 @@
 class Debug
 {
   public:
-    void DebugNum(char label[], int num);
-    void DebugNum2(char label[], int num1, int num2);
+    void DebugNum(char label[], byte num);
+    void DebugNum2(char label[], byte num1, byte num2);
   private:
 };
 
-void Debug::DebugNum(char label[], int num)
+void Debug::DebugNum(char label[], byte num)
 {
   Serial.print(label);
   Serial.println(num);
 }
 
-void Debug::DebugNum2(char label[], int num1, int num2)
+void Debug::DebugNum2(char label[], byte num1, byte num2)
 {
   Serial.print(label);
   Serial.print(num1);
